+++ src/cli/ocat/commands/run.ts
@@ -1,9 +1,9 @@
 import { execute } from "../../../core/index.js";
-import { OcatError } from "../../../shared/manager/errors/coreErrors.js";
-
+import { ExtensionError } from "../../../shared/manager/errors/io/extension.js";
 import chalk from "chalk";
 
 import fs from "fs";
+import { FileDoesntExistError } from "../../../shared/manager/errors/io/file.js";
 
 export interface RunCommandOptions {
     force: boolean;
@@ -11,14 +11,19 @@ export interface RunCommandOptions {
 
 export function runfile(file: string, options: RunCommandOptions) {
     if (!file.endsWith(".ocat") && !options.force) { 
-        new OcatError("File must be a .ocat file. Use -f to force execution with other extensions")
-            .throw();
-        process.exit(1);
+		new ExtensionError(
+			"File must be a .ocat file. Use -f to force execution with other extensions"
+		).throw();
     }
 
     if (options.force) {
         console.log(chalk.blue(`Running in force mode`));
-    }
+	}
+	
+	if(!fs.existsSync(file)) {
+		new FileDoesntExistError(`File ${file} doesn't exist`)
+			.throw();
+	}
 
     const fileText = fs.readFileSync(file, "utf8");