+++ src/shared/manager/baseManager.ts
@@ -12,9 +12,14 @@ export class OcatManager {
 	}
 
 	toString(line: number | undefined = undefined): string {
-		return `${this.color.bold(this.name)} ${
-			line && chalk.gray(`at line ${line}`)
+		return `${this.color.bold(this.name)}${
+			line ? chalk.gray(` at line ${line}`) : ""
 		}: ${this.color.italic(this.message)}`;
+	}
+
+	setMessage(message: string): this {
+		this.message = message;
+		return this;
 	}
 
 	throw(line: number | undefined = undefined): never {