+++ src/shared/manager/errors/io/extension.ts
@@ -1,0 +1,8 @@
+import { OcatError } from "../coreErrors.js";
+
+export class ExtensionError extends OcatError {
+	constructor(message: string) {
+		super("ExtensionError");
+		this.setMessage(message);
+	}
+}