+++ src/shared/manager/errors/io/file.ts
@@ -1,0 +1,8 @@
+import { OcatError } from "../coreErrors.js";
+
+export class FileDoesntExistError extends OcatError {
+	constructor(message: string) {
+		super("FileDoesn'tExistError");
+		this.setMessage(message);
+	}
+}