+++ src/shared/manager/errors/semantic/undefined/declared.ts
@@ -4,7 +4,7 @@ import { OcatError } from "../../coreErrors.js";
 export class AlreadyDeclaredError extends OcatError {
 	constructor(varName: string, type: ContextType) {
 		super(`Already declared ${type.toLowerCase()}`);
-		this.message = `${type} ${varName} is already declared`;
+		this.setMessage(`${type} ${varName} is already declared`);
 	}
 }
 
@@ -23,6 +23,6 @@ export class AlreadyDeclaredFunctionError extends AlreadyDeclaredError {
 export class CantModifyConstError extends OcatError {
 	constructor(varName: string) {
 		super(`Can't modify const variable`);
-		this.message = `Variable ${varName} is const and can't be modified`;
+		this.setMessage(`Variable ${varName} is const and can't be modified`);
 	}
 }