+++ src/shared/manager/errors/syntax/syntaxErrors.ts
@@ -3,6 +3,6 @@ import { OcatError } from "../coreErrors.js";
 export class SyntaxError extends OcatError {
 	constructor(message: string) {
 		super("SyntaxError");
-		this.message = message;
+		this.setMessage(message);
 	}
 }